moved challenge list out of container
